+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5] - 2026-07-02
+
+### Added
+- Added VeinMiner crop replant compatibility.
+
+### Changed
+- Crop harvest and replant now uses left-click block breaking instead of right-click interaction.
+
+## [0.4] - 2026-07-02
+
+### Added
+- Merged NextGenNether functionality into HappyFarmer.
+- Added Shulker Box recipe: Chest surrounded by Netherrack.
+- Added Ender Chest recipe: Chest surrounded by Blackstone.
+- Added Netherrack furnace fuel with 300 ticks of burn time.
+
 ## [0.3] - 2026-07-02
 
 ### Added